// Ecommerce.Frontend/Ecommerce/src/app/core/models/auth.models.ts
export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  fullName: string;
  email: string;
  password: string;
  phoneNumber: string;
  role?: string;
}

export interface UserResponse {
  userId: number;
  email: string;
  fullName: string;
  role: string;
  token: string;
}